Use let+const

+++ src/asar.js
@@ -1,14 +1,14 @@
 'use strict';
-var fs = require('fs');
-var path = require('path');
-var os = require('os');
-var minimatch = require('minimatch');
-var mkdirp = require('mkdirp');
+const fs = require('fs');
+const path = require('path');
+const os = require('os');
+const minimatch = require('minimatch');
+const mkdirp = require('mkdirp');
 
-var Filesystem = require('./filesystem');
-var disk = require('./disk');
-var crawlFilesystem = require('./crawlfs');
-var createSnapshot = require('./snapshot');
+const Filesystem = require('./filesystem');
+const disk = require('./disk');
+const crawlFilesystem = require('./crawlfs');
+const createSnapshot = require('./snapshot');
 
 // Return whether or not a directory should be excluded from packing due to
 // "--unpack-dir" option
@@ -16,8 +16,8 @@ var createSnapshot = require('./snapshot');
 // @param {string} path - diretory path to check
 // @param {string} pattern - literal prefix [for backward compatibility] or glob pattern
 //
-var isUnpackDir = function(path, pattern) {
-  return path.indexOf(pattern) === 0 || minimatch(path, pattern);
+const isUnpackDir = function(path, pattern) {
+  return path.startsWith(pattern) || minimatch(path, pattern);
 };
 
 module.exports.createPackage = function(src, dest, callback) {
@@ -25,8 +25,7 @@ module.exports.createPackage = function(src, dest, callback) {
 };
 
 module.exports.createPackageWithOptions = function(src, dest, options, callback) {
-  var dot = options.dot;
-  if (dot === undefined) { dot = true; }
+  const dot = typeof options.dot === 'undefined' ? true : options.dot;
 
   return crawlFilesystem(src, { dot: dot }, function(error, filenames, metadata) {
     if (error) { return callback(error); }
@@ -45,39 +44,40 @@ callback: The callback function. Accepts (err).
 */
 module.exports.createPackageFromFiles = function(src, dest, filenames, metadata, options, callback) {
   if (typeof metadata === 'undefined' || metadata === null) { metadata = {}; }
-  var filesystem = new Filesystem(src);
-  var files = [];
+  const filesystem = new Filesystem(src);
+  const files = [];
 
+  let filenamesSorted = [];
   if (options.ordering) {
-    var orderingFiles = fs.readFileSync(options.ordering).toString().split('\n').map(function(line) {
-      if (line.indexOf(':') !== -1) { line = line.split(':').pop(); }
+    const orderingFiles = fs.readFileSync(options.ordering).toString().split('\n').map(function(line) {
+      if (line.includes(':')) { line = line.split(':').pop(); }
       line = line.trim();
-      if (line[0] === '/') { line = line.slice(1); }
+      if (line.startsWith('/')) { line = line.slice(1); }
       return line;
     });
 
-    var ordering = [];
-    for (var file of orderingFiles) {
-      var pathComponents = file.split(path.sep);
-      var str = src;
-      for (var pathComponent of pathComponents) {
+    const ordering = [];
+    for (const file of orderingFiles) {
+      const pathComponents = file.split(path.sep);
+      let str = src;
+      for (const pathComponent of pathComponents) {
         str = path.join(str, pathComponent);
         ordering.push(str);
       }
     }
 
-    var filenamesSorted = [];
-    var missing = 0;
-    var total = filenames.length;
+
+    const missing = 0;
+    const total = filenames.length;
 
     for (file of ordering) {
-      if (filenamesSorted.indexOf(file) === -1 && filenames.indexOf(file) !== -1) {
+      if (!filenamesSorted.includes(file) && filenames.includes(file)) {
         filenamesSorted.push(file);
       }
     }
 
     for (file of filenames) {
-      if (filenamesSorted.indexOf(file) === -1) {
+      if (!filenamesSorted.includes(file)) {
         filenamesSorted.push(file);
         missing += 1;
       }
@@ -85,26 +85,26 @@ module.exports.createPackageFromFiles = function(src, dest, filenames, metadata,
 
     console.log(`Ordering file has ${((total - missing) / total) * 100}% coverage.`);
   } else {
-    var filenamesSorted = filenames;
+    filenamesSorted = filenames;
   }
 
-  var handleFile = function(filename, done) {
-    var file = metadata[filename];
+  const handleFile = function(filename, done) {
+    let file = metadata[filename];
+    let type;
     if (!file) {
-      var stat = fs.lstatSync(filename);
-      if (stat.isDirectory()) { var type = 'directory'; }
-      if (stat.isFile()) { var type = 'file'; }
-      if (stat.isSymbolicLink()) { var type = 'link'; }
+      const stat = fs.lstatSync(filename);
+      if (stat.isDirectory()) { type = 'directory'; }
+      if (stat.isFile()) { type = 'file'; }
+      if (stat.isSymbolicLink()) { type = 'link'; }
       file = {stat, type};
     }
 
+    let shouldUnpack;
     switch (file.type) {
       case 'directory':
-        var shouldUnpack =
-          options.unpackDir ?
-            isUnpackDir(path.relative(src, filename), options.unpackDir)
-          :
-            false;
+        shouldUnpack = options.unpackDir
+          ? isUnpackDir(path.relative(src, filename), options.unpackDir)
+          : false;
         filesystem.insertDirectory(filename, shouldUnpack);
         break;
       case 'file':
@@ -113,7 +113,7 @@ module.exports.createPackageFromFiles = function(src, dest, filenames, metadata,
           shouldUnpack = minimatch(filename, options.unpack, {matchBase: true});
         }
         if (!shouldUnpack && options.unpackDir) {
-          var dirName = path.relative(src, path.dirname(filename));
+          const dirName = path.relative(src, path.dirname(filename));
           shouldUnpack = isUnpackDir(dirName, options.unpackDir);
         }
         files.push({filename: filename, unpack: shouldUnpack});
@@ -127,7 +127,7 @@ module.exports.createPackageFromFiles = function(src, dest, filenames, metadata,
     return process.nextTick(done);
   };
 
-  var insertsDone = function() {
+  const insertsDone = function() {
     return mkdirp(path.dirname(dest), function(error) {
       if (error) { return callback(error); }
       return disk.writeFilesystem(dest, filesystem, files, metadata, function(error) {
@@ -141,9 +141,9 @@ module.exports.createPackageFromFiles = function(src, dest, filenames, metadata,
     });
   };
 
-  var names = filenamesSorted.slice();
+  const names = filenamesSorted.slice();
 
-  var next = function(name) {
+  const next = function(name) {
     if (!name) { return insertsDone(); }
 
     return handleFile(name, function() {
@@ -155,7 +155,7 @@ module.exports.createPackageFromFiles = function(src, dest, filenames, metadata,
 };
 
 module.exports.statFile = function(archive, filename, followLinks) {
-  var filesystem = disk.readFilesystemSync(archive);
+  const filesystem = disk.readFilesystemSync(archive);
   return filesystem.getFile(filename, followLinks);
 };
 
@@ -164,47 +164,45 @@ module.exports.listPackage = function(archive) {
 };
 
 module.exports.extractFile = function(archive, filename) {
-  var filesystem = disk.readFilesystemSync(archive);
+  const filesystem = disk.readFilesystemSync(archive);
   return disk.readFileSync(filesystem, filename, filesystem.getFile(filename));
 };
 
 module.exports.extractAll = function(archive, dest) {
-  var destFilename;
-  var file;
-  var linkSrcPath;
-  var linkDestPath;
-  var relativePath;
-  var linkTo;
-  var content;
-  var filesystem = disk.readFilesystemSync(archive);
-  var filenames = filesystem.listFiles();
+  const filesystem = disk.readFilesystemSync(archive);
+  const filenames = filesystem.listFiles();
 
   // under windows just extract links as regular files
-  var followLinks = process.platform === 'win32';
+  const followLinks = process.platform === 'win32';
 
   // create destination directory
   mkdirp.sync(dest);
 
-  return filenames.map((filename) =>
-    (filename = filename.substr(1),  // get rid of leading slash
-    destFilename = path.join(dest, filename),
-    file = filesystem.getFile(filename, followLinks),
-    file.files ?
+  return filenames.map((filename) => {
+
+    filename = filename.substr(1);  // get rid of leading slash
+    const destFilename = path.join(dest, filename);
+    const file = filesystem.getFile(filename, followLinks);
+    if (file.files) {
       // it's a directory, create it and continue with the next entry
-      mkdirp.sync(destFilename)
-    : file.link ?
+      mkdirp.sync(destFilename);
+    } else if(file.link) {
       // it's a symlink, create a symlink
-      (linkSrcPath = path.dirname(path.join(dest, file.link)),
-      linkDestPath = path.dirname(destFilename),
-      relativePath = path.relative(linkDestPath, linkSrcPath),
+      const linkSrcPath = path.dirname(path.join(dest, file.link))
+      const linkDestPath = path.dirname(destFilename);
+      const relativePath = path.relative(linkDestPath, linkSrcPath);
       // try to delete output file, because we can't overwrite a link
-      (() => { try {
-        result.push(fs.unlinkSync(destFilename));
-      } catch (error) {} })(),
-      linkTo = path.join(relativePath, path.basename(file.link)),
-      fs.symlinkSync(linkTo, destFilename))
-    :
+      (() => {
+        try {
+          result.push(fs.unlinkSync(destFilename));
+        } catch (error) {}
+      })();
+      const linkTo = path.join(relativePath, path.basename(file.link));
+      fs.symlinkSync(linkTo, destFilename);
+    } else {
       // it's a file, extract it
-      (content = disk.readFileSync(filesystem, filename, file),
-      fs.writeFileSync(destFilename, content))));
+      const content = disk.readFileSync(filesystem, filename, file);
+      fs.writeFileSync(destFilename, content);
+    }
+  });
 };
